+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Features:
 - feed: change highlight color of an item, if it has a reading progress
 - main: Navigation moved from the header into the footer
 - db: allow a folder as node name in the configuration to fetch all feeds that are directly in this folder
+- login: redirect to requested page after login
 
 Fixes: